+++ src/stream_ml/pytorch/_multi.py
@@ -22,7 +22,7 @@ from stream_ml.pytorch.typing import Array, NNModel
 from stream_ml.pytorch.utils import names_intersect
 
 if TYPE_CHECKING:
-    from stream_ml.core.data import Data
+    from stream_ml.pytorch import Data
 
     Self = TypeVar("Self", bound="MixtureModel")
