+++ src/stream_ml/pytorch/builtin/__init__.py
@@ -21,7 +21,6 @@ from dataclasses import field, make_dataclass
 import torch as xp
 
 from stream_ml.core.builtin._uniform import Uniform as CoreUniform
-from stream_ml.core.typing import ArrayNamespace
 
 from stream_ml.pytorch._base import ModelBase
 from stream_ml.pytorch.builtin import compat
@@ -36,7 +35,7 @@ from stream_ml.pytorch.builtin._skewnorm import SkewNormal
 from stream_ml.pytorch.builtin._sloped import Sloped
 from stream_ml.pytorch.builtin._truncnorm import TruncatedNormal
 from stream_ml.pytorch.builtin._truncskewnorm import TruncatedSkewNormal
-from stream_ml.pytorch.typing import Array, NNModel
+from stream_ml.pytorch.typing import Array, ArrayNamespace, NNModel
 
 Uniform = make_dataclass(
     "Uniform",
