+++ src/stream_ml/pytorch/builtin/_exponential.py
@@ -7,10 +7,9 @@ from dataclasses import field, make_dataclass
 import torch as xp
 
 from stream_ml.core.builtin._exponential import Exponential as CoreExponential
-from stream_ml.core.typing import ArrayNamespace
 
 from stream_ml.pytorch._base import ModelBase
-from stream_ml.pytorch.typing import Array, NNModel
+from stream_ml.pytorch.typing import Array, ArrayNamespace, NNModel
 
 Exponential = make_dataclass(
     "Exponential",
