+++ src/stream_ml/pytorch/builtin/_multinormal.py
@@ -13,9 +13,9 @@ from torch.distributions import MultivariateNormal as TorchMultivariateNormal
 from stream_ml.pytorch._base import ModelBase
 
 if TYPE_CHECKING:
-    from stream_ml.core.data import Data
     from stream_ml.core.params import Params
 
+    from stream_ml.pytorch import Data
     from stream_ml.pytorch.typing import Array
