+++ src/stream_ml/pytorch/builtin/_skewnorm.py
@@ -11,9 +11,9 @@ from stream_ml.pytorch._base import ModelBase
 from stream_ml.pytorch.builtin._stats.skewnorm import logpdf
 
 if TYPE_CHECKING:
-    from stream_ml.core.data import Data
     from stream_ml.core.params import Params
 
+    from stream_ml.pytorch import Data
     from stream_ml.pytorch.typing import Array
