+++ src/stream_ml/pytorch/builtin/_truncskewnorm.py
@@ -12,9 +12,9 @@ from stream_ml.pytorch.builtin._stats.truncskewnorm import logpdf
 from stream_ml.pytorch.builtin._truncnorm import TruncatedNormal
 
 if TYPE_CHECKING:
-    from stream_ml.core.data import Data
     from stream_ml.core.params import Params
 
+    from stream_ml.pytorch import Data
     from stream_ml.pytorch.typing import Array
