+++ src/stream_ml/pytorch/builtin/compat/flow.py
@@ -16,9 +16,9 @@ from stream_ml.pytorch._base import ModelBase
 from stream_ml.pytorch.utils import names_intersect
 
 if TYPE_CHECKING:
-    from stream_ml.core.data import Data
     from stream_ml.core.params import Params
 
+    from stream_ml.pytorch import Data
     from stream_ml.pytorch.typing import Array
