+++ src/stream_ml/pytorch/typing.py
@@ -1,6 +1,8 @@
 """Pytorch type hints."""
 
+__all__ = ["Array", "NNModel", "ArrayNamespace"]
+
 from torch import Tensor as Array
 from torch.nn import Module as NNModel
 
-__all__ = ["Array", "NNModel"]
+from stream_ml.core.typing import ArrayNamespace
